make it actually work
right now I have an imgBytes behavior that contains the ByteString of the image and when I update that I send it to the js running in the browser to display it.
It's still a bit impractical since I save and load the image before processing and have to press the draw/load buttons since on the js side it happens with callbacks but it works and I was able to make a recursive Atul.